♻️ refactor : 채팅 멘토링 공간좌표 적용

+++ src/main/java/roomit/main/domain/workplace/dto/response/WorkplaceAllResponse.java
@@ -9,8 +9,8 @@ public record WorkplaceAllResponse(
         String imageUrl,
         Double stars,
         Long reviewCount,
-        BigDecimal positionLat,
-        BigDecimal positionLon,
+        Double positionLat,
+        Double positionLon,
         Double distance
 ) {}
